updating test util and implementing more tests

+++ src/server/api/impressions/destroy.js
@@ -2,10 +2,15 @@ import Impression from '../../models/impression'
 
 const route = async (req, res) => {
 
-  const impression = await Impression.where({
-    id: req.params.id
+  const impression = await Impression.query(qb => {
+    qb.where('visit_id', req.visit.get('id'))
+    qb.where('id', req.params.id)
   }).fetch({
     transacting: req.trx
+  })
+
+  if(!impression) return res.status(404).json({
+    message: 'Unable to load impression'
   })
 
   await impression.destroy({
